Make Command initializer public
Public Command initializer allows any command to be called. Public perform function in Connection class allows any command to be sent and parsed as long as the caller can provide a Codable model that is expected to be received as a response.

Add IDsResponse model to handle list of ids for a given key.